Started ChipStack
The ChipStack script for stacking the chips placed on each bet has been started. It currently contains a float value for the height at which the next chip should be placed and an array of gameObjects to act as the 'base' of each chip.
Next, each gameObject for the chips themselves need to be added as well as utilising the position of each object in the array.