Raise/Fold/Call Buttons
Added in each object for the 3 sets of buttons. Plugged in functionality and adjusted code base to work with new design for buttons use (using the card objects and having a small display). Display is a small chalk board underneath which gives details on current phase and what to do.

Generally polished referencing between scripts and the functionality used to call between them.

Also removed all old scripts due to errors occurring when editing scripts that are refenced in both Old and New.