Fixed Camera Issues
Now possible to move from the table to the end screen. Check the Rail Script for all new additions. Unity Events have been provided for the necessary moments in game. Rails can be reset. Also added methods for the look at method as well as a list of targets in the hierarchy.