Debugging Betting Panel
Fixed some of the bugs with the panel. fixed logic to prevent constant addition of chips to the players "stash" along with other small changes related to.

Current bugs left are player "stash" not deducting chips bet on other piles & glitch where switching between options eventually defaults panel to show same info for all 3, but name for which selected is updated