Opponent Call&Reset&Jalp
Added in reset methods fro resetting the game to default for reppeat playing.

Added in AI decisions and code for calling the players raise & added + plugged in the logic for this and the AI raising.

Added methods for opening/closing the help screen from a button press, and added said buttons to the pages which work