Dealer label and betting panel selection UI
Added a time delay for Dealer updating the text field, to allow time for the flip animation.

Also added some smaller UI additions for better UX, such as info on the start menu about the help button.

Also added in white, circular markers to indicate the selected chip size on the betting panel.