Plugged in animations and audio. Remade some audio and animation clips. Found some bugs.
Added all animations except for the chips as they would look silly.
Added more transitions to Card and BettingPanel animations so that they can be cycled through for each round.

Implemented AudioManager into main scene and updated the sounds inside it. Had to remake some as they were inaudible over the music. Created simple click and invalidBet sounds for button presses for more use feedback.

Found UI bugs with the AI chips and the player InfoPanel not resetting on starting a subsequent game.
Also found a minor audio bug that is caused by how/when an event is called, but it's minor.

Fixed small error in an event call in the raiseFold phase.